interface support of framework
