+++ cautils/opapolicy/datastructures.go
@@ -1,7 +1,6 @@
 package opapolicy
 
 import (
-	"path/filepath"
 	"time"
 
 	armotypes "github.com/armosec/kubescape/cautils/armotypes"
@@ -149,44 +148,4 @@ type PolicyNotification struct {
 type PolicyIdentifier struct {
 	Kind NotificationPolicyKind `json:"kind"`
 	Name string                 `json:"name"`
-}
-
-type ScanInfo struct {
-	PolicyIdentifier   PolicyIdentifier
-	Format             string
-	Output             string
-	ExcludedNamespaces string
-	InputPatterns      []string
-	Silent             bool
-}
-
-func (scanInfo *ScanInfo) Init() {
-	// scanInfo.setSilentMode()
-	scanInfo.setOutputFile()
-
-}
-
-func (scanInfo *ScanInfo) setSilentMode() {
-	if scanInfo.Format == "json" || scanInfo.Format == "junit" {
-		scanInfo.Silent = true
-	}
-	if scanInfo.Output != "" {
-		scanInfo.Silent = true
-	}
-}
-
-func (scanInfo *ScanInfo) setOutputFile() {
-	if scanInfo.Output == "" {
-		return
-	}
-	if scanInfo.Format == "json" {
-		if filepath.Ext(scanInfo.Output) != "json" {
-			scanInfo.Output += ".json"
-		}
-	}
-	if scanInfo.Format == "junit" {
-		if filepath.Ext(scanInfo.Output) != "xml" {
-			scanInfo.Output += ".xml"
-		}
-	}
 }
