+++ policyhandler/handlenotification.go
@@ -4,6 +4,7 @@ import (
 	"fmt"
 
 	"github.com/armosec/kubescape/cautils"
+	"github.com/armosec/kubescape/cautils/getter"
 
 	"github.com/armosec/kubescape/cautils/k8sinterface"
 
@@ -25,13 +26,13 @@ func NewPolicyHandler(processPolicy *chan *cautils.OPASessionObj, k8s *k8sinterf
 	}
 }
 
-func (policyHandler *PolicyHandler) HandleNotificationRequest(notification *opapolicy.PolicyNotification, scanInfo *opapolicy.ScanInfo) error {
+func (policyHandler *PolicyHandler) HandleNotificationRequest(notification *opapolicy.PolicyNotification, scanInfo *cautils.ScanInfo) error {
 	opaSessionObj := cautils.NewOPASessionObj(nil, nil)
 	// validate notification
 	// TODO
 
 	// get policies
-	frameworks, err := policyHandler.getPolicies(notification)
+	frameworks, err := policyHandler.getPolicies(notification, scanInfo.PolicyGetter)
 	if err != nil {
 		return err
 	}
@@ -54,12 +55,12 @@ func (policyHandler *PolicyHandler) HandleNotificationRequest(notification *opap
 	return nil
 }
 
-func (policyHandler *PolicyHandler) getPolicies(notification *opapolicy.PolicyNotification) ([]opapolicy.Framework, error) {
+func (policyHandler *PolicyHandler) getPolicies(notification *opapolicy.PolicyNotification, policyGetter getter.IPolicyGetter) ([]opapolicy.Framework, error) {
 
-	cautils.ProgressTextDisplay("Downloading framework definitions")
+	cautils.ProgressTextDisplay("Downloading/Loading framework definitions")
 
 	// TODO - support load policies from local file
-	frameworks, err := policyHandler.GetPoliciesFromBackend(notification)
+	frameworks, err := policyHandler.GetPoliciesFromBackend(notification, policyGetter)
 	if err != nil {
 		return frameworks, err
 	}
@@ -68,12 +69,12 @@ func (policyHandler *PolicyHandler) getPolicies(notification *opapolicy.PolicyNo
 		err := fmt.Errorf("could not download any policies, please check previous logs")
 		return frameworks, err
 	}
-	cautils.SuccessTextDisplay("Downloaded framework")
+	cautils.SuccessTextDisplay("Downloaded/Loaded framework")
 
 	return frameworks, nil
 }
 
-func (policyHandler *PolicyHandler) getResources(notification *opapolicy.PolicyNotification, opaSessionObj *cautils.OPASessionObj, scanInfo *opapolicy.ScanInfo) (*cautils.K8SResources, error) {
+func (policyHandler *PolicyHandler) getResources(notification *opapolicy.PolicyNotification, opaSessionObj *cautils.OPASessionObj, scanInfo *cautils.ScanInfo) (*cautils.K8SResources, error) {
 	var k8sResources *cautils.K8SResources
 	var err error
 	if len(scanInfo.InputPatterns) > 0 {
